pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "climatter"
version = "v0.1.0"
description = "Remember what matters - A CLI tool to track and countdown to your important days"
readme = "README.md"  # Optional, only if you create one
requires-python = ">=3.9"
dependencies = [
    "rich==14.3.2",
    "pyyaml==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pylint==4.0.4",
]

[tool.setuptools.package-data]
climatter = ["default_config/*.yaml"]

[project.scripts]
climatter = "climatter.main:main"

[tool.black]
line-length = 80
target-version = ['py39']
include = '\.pyi?$'
